requirements.txt
Phidget22==1.7.20211005
requests==2.32.0
pyusb==1.0.2
gpiozero==1.5.1
pigpio==1.78
paho-mqtt==2.1.0
